NH-3952: more cleanup.
